move encryption test
